requirements.txt
Django==1.7.1
Markdown==2.5.1
MySQL-python==1.2.5
Pillow==2.6.1
argparse==1.2.1
django-bower==5.0.1
django-crispy-forms==1.4.0
django-filter==0.8
django-gravatar2==1.1.4
django-recaptcha==1.0.2
djangorestframework==2.4.4
python-openid==2.2.5
requests==2.4.3
requests-oauthlib==0.4.2
six==1.8.0
wsgiref==0.1.2
